Reorganizing lib structure
